Edit where('exists') clause condition
  * Now, where('exists') first orders by the default field, and then orders
by the existsField. This makes sure that ordering is primarily based on
the default field, and in most situations, the existsField simply acts
as filter, or at worst a tie-breaker.
  * Set undefined to be assigned to the default field, in the
DBQuery child classes. It is explicitly set as a particular field, not
__name__, as firebase is weird about ordering after __name__. Any unknown
fields return an error, and require manually adding to the class.
  * Fixed fieldName bug in inequality clause auto orderBy adder.
Also made corresponding changes in MockDBFetch