Make LoginModal a persistent component
  * It has been given its own context, so it can be opened from any
component in the application(i.e. triggering a login).
  * Hence we removed this loginModal from the Navbar, and moved it,
along with NotificationSnackbar, into PersistentComponents component.
This component symbolizes stuff which can be called from any part of
the application.
  * Also changed the firstMount logic to trigger when the open state
changes. This eliminates the need to have the required logic for the
component outside the compnent, as was earlier
  * Renamed the context within the LoginModalPages, to support this new
context for calling LoginModal from anywhere in the app