Add addressList to userData
  * AddressList is an object of Address instances. Each address has its
own uid along with its contents. It is intended that the Address instance
be handled by the using component for its data, and also passed into
AddressListInstance.remove() method, so that it can delete it with the uid
set internally as a private key. In order to store these address instances
in the cloud DB, we also created a `objectifiedContents` getter function,
along with a `createFromObjects` static function to convert from DB.
  * Also fixed minor bug about cloning. The handler classes were
explicitly cloning before running the `add`/`remove`/`toggle` functions.
This is unneccessary, as the cloning happens internally anyway.