Add AddressSelector component, and modify ScrollingSelect styles
  * We can see all the addresses and select them also. Providing the
onSelect prop activates selected Address highlighting. Right now we can
see the addresses, and select them. In the future we could manipulate
these addesses by editing, deleting and adding new addresses, from this
component.
  * In order to support selecting functionality, a find method was added
to AddressList, which uses the addressID to find retrieve the address.
  * ScrollingSelect was modified such that the Stack had an alignItems of
center. This allows the extension of this component, while making use of
the flexbox ability to stretch all items as large as the biggest item.
This provides visual consistency. But in the case of any customization,
this can be overriden, by simply providing a Box as the ScrollingSelect
child(which becomes the inputLabel), and then making any layout changes
within that Box.