Refactor CartPage to use Suspense, and Store data in cache
  * Using the suspense component, we handled loading and error states.
  * The data stored in cache only updates if there is an unknown new
product to be listed. This improves page reactivity, and eliminates
redundant remote fetches.
  * Got rid of wrapper contents() getter fn in CartHandler. It helps us
more when we use the object itself, cause its easier to access the
individual items. If we need an array to list, we can just create it at
the consuming end.
  * Included productID in the CartItem data. This helps us have an easy
access to the individual items, and also eliminates the need to share the
generateProductID fn, which is now scoped as a base class constructor
static methods.