Add checkout path
  * The checkout function is provided via UserContext. It navigates to
the checkout page, and provides the required cart via react-router
state/useLocation. Due to the nature of this feature, we need to serialize
and deserialize the cart(Existing functions do the trick). Also, it
reroutes any manual access of the checkout page, to the home page.
  * The GUI is the checkout page, which shows the orderTotal, and
provides a form for delivery options. The data for the page is fetched
using the CheckoutDataHandler class, which provides a neat interface for
fetching and accessing the data, instead of having it all inside the
react component.
  * This checkout functionality was added to CartPage and ProductPage.
  * Also modified AddressSelector to accept the name prop, and modified
the return value, so that we can use generic handler methods, to update
react state.
  * NOTE: At this point the data is only collected, and is not sent to
the backend.