Refactor DB Manipulators to use docPaths instead of docNames
  * Having docPaths instead of docNames allows us to dynamically provide
multiple documentIDs in the form of docNameAliases. The order in which
these aliases construct the final path is determined at the time of
Database creation, so we do not have to worry about the order in which we
provide these keys in the calling object.
  * We moved this extendedMethods into the base class definition. The
additional functionality we provided with the extendedMethods was
incorporated into the base class, and this simplifies the internal logic.
  * This improved pattern allows us to access nested documents such as
subcollections, upto any degree of nesting.
  * Also fixed typing error which mixed up the DB_NAME for Users and
Game DBs.
  * Corresponding migration to docPath was also made in DBQueryClasses.
This allows us to run queries also, on nested collections.
  * MockDBFetch was also refactored to accomodate these changes