Connect backend to checkout page deliveryForm
  * Handles the data submitting, and shows the user a success or failure
message. Submit button is disabled during submitting.
  * Also added a totalPrice counter, which takes into account the
additional charges of delivery options. This price is calculated
dynamically, depending on each delivery option, and its price. The backend
doesn't receive the price, only the options. This prevents tampering with
the cost. Right now the displayed price for each delivery option is
hardcoded; but that can be easily set to load when the module is loaded.
  * Modified the NotificationSnackbarContext to be initialized with the
correct context values. This helps with code suggestions in vscode.
  * The addPrices fn now will also function as an identity function when
provided with a single argument.
  * Moved CheckoutList into its own file to keep things clean.