Add auth check to modify userData
  * The loginModal is shown if the user isn't logged in, and tries to
attempt a privileged operation(addTo/removeFrom Cart/Wishlist etc.)
  * In order to wrap the UserDataHandlers with auth checks, the
showLoginModal fn was passed as an argument to the constructor. The
constructor in turn was refactored to accept an object as an argument,
so that things are more clear.
  * The order of context providers was switched up, so that UserContext
has access to the LoginModalContext(so that it can be called)
  * Added default keys to createContext in LoginModal context to help in
VSCode code completion.
  * Fixed bug where the userData status was the same even after logging
out. Now it is emptied on logout.