Abstract out UserContextHandlerClasses, into generator class.
  * The main responsiblitiy of UserContextHandlerClasses was to extend
the base functionality with the UserData Context features like state.
Earlier we were using a dedicated extender class for each object, but
this was quickly getting out of hand with a lot of code duplication. The
core functionality of all extender classes was the same; so we have
abstracted it into a single proxy generator. Earlier we tried to implement
this using Action Filters, as implemented in ASP.NET, but it turns out we
can use the super useful JS feature of proxy. The proxy wraps the method
calls with auth checks and state updates, as specified at the time we
initiate/extend the base class, by accessing the react state and
loginContext, which we provide at the time of initializing Proxy
generator. The contextHandler/proxy generator is memoized as a
performance optimization.