Refactor CheckoutPage to use useResource custom hook
  * A lot of code was being duplicated in each component that used
Suspense boundaries. We extracted this into a custom hook useResource,
which takes in an async function that returns the data of the resource.
The hook itself returns the resource to be used in our react components.
It is initialized with an empty promise, and then is immediately updated
in the next render, with the return value of the dataFetch fn.