Refactor CheckoutPageHelper to make use of class features
  * Instead of handling all the data in one place, the CheckoutDataHandler
is the main class that handles the overall data for the checkout page.
The individual pieces of data are handled by smaller classes. Also the
features like getters and setters were used to get rid of spaghetti
functions all over the place. Now everything looks much more organized.
  * CheckoutOrderItem handles the data for the individual items that are
to be checked out. It handles calling the backend fns, so that its
wrapping classes don't have to worry about it. At this point in time, it
fetches the data for each item, without any sort of caching.