Refactor CartPage to make use of class features, and useResource hook
  * Using a dedicated DataHandler class, removes the data handling logic
from the react component, so that it can focus solely on the UI stuff.
Additionally, useResource cleans up a lot of the boilerplate regarding
resource creation.
  * The productDataCache was abstracted into its own helper class.
Earlier CartProductDataCache would provide a CartItemDataHandler; which
is a highly coupled solution. This was eliminated. The data from the
cache can be used in multiple places, by generating data handlers
specific to each component.
  * The old CartItemDataHandler was deleted, along with the CartProductCache.
These did not make use of class features like getters and setters.