Add ProductPage that uses helper classes and useResource
  * The component is slightly convoluted as there are multiple
subcomponents that depend on the same productDataResource. So instead of
having a Suspense-Error Boundary for each resource consuming component,
we bundled it together as a ProductPageFallback, which is composed of all
these individual fallback components.
  * The datahandler classes handle fetching and providing data as an
object.
  * Deleted the ProductPageItem from DBDataConverter, since it is now
located next to the consuming component i.e. ProductPage.
  * CounterButtons was extracted into its own component that can be
reused elsewhere.
  * Added resourcePropType to SuspenseHelpers(prevents repeating propType
for resource), and modified the promise getter to return the actual promise
that returns the data, instead of the suspender.