Move deliveryOptions for checkout into database
  * Earlier the delivery options were hardcoded in the code. Now they
are contained in a separate orders metadoc in a metadata collection. To
handle this metadatabase, a new DBManipulator called Metadatabase was added.
  * A dedicated OrdersMetadata class handles the metadata provided by
the orders document in metadata collection.
  * Added firestore rules to access metadata
  * CheckoutPage now handles two resources, one for the itemData, and
one for the formData.
  * Renamed CheckoutDataHandler to CheckoutItemsHandler, so that the
former name can be used for the new dataHandler, while the latter is
included in the new dataHandler.
  * Since the checkoutPage now handles a lot of data, a custom hook
called useDataHandler was used. This creates a parallel object, for the
react component, which handles all of the data fetching and store. It
accepts both class instances, as well as class definitions, with static
methods(in case of no need for maintaining state).
  * DeliveryForm now uses suspense internally, and renders a fallback
UI. Also, delivery form has its own dedicated dataHandler, since it is
quite convoluted. Further, we split up the data collected in the form
into address and deliveryOptions(as opposed to both being bundled together
under the latter name). This is because the deliveryOptions are booleans
and have a very specific context, in connection with the ordersMetadata
doc.
  * OrderClass was modified to separate address and deliveryOptions.