Hookup SearchBar/SearchBox to algolia search, and fix useDataHandler bug
  * The SearchBar/Box was moved out of the navbar component into its own
folder; with subcomponents and helpers.
  * The SearchBox uses an MUI autocomplete to render the options, while
the rest of the UI is the same as before(using an InputBase component
from MUI). As soon as an option is clicked the user is routed to the
corresponding product page.
  * The options are fetched dynamically via algolia. After the results
are fetched, corresponding images are fetched from firebase, while
making sure of a fallback UI using React Suspense. Only required data is
fetched from the backend to reduce network latency.
  * Fixed bug in useDataHandler. It was checking for a dispose method on
the instance of the object instead of its prototype, where the method
actually lies. Now it's fixed by using `in` instead of `Object.hasOwn`